+++ matrix.go
@@ -12,6 +12,12 @@ func (m Mat4) MulVec4(v Vec4) Vec4 {
 		m.m[3][0]*v.x + m.m[3][1]*v.y + m.m[3][2]*v.z + m.m[3][3]*v.w,
 	}
 }
+
+// Return an Identity Matrix
+// | 1  0  0  0 |
+// | 0  1  0  0 |
+// | 0  0  1  0 |
+// | 0  0  0  0 |
 func Mat4Identity() Mat4 {
 	return Mat4{[4][4]float64{
 		{1, 0, 0, 0},
@@ -21,10 +27,28 @@ func Mat4Identity() Mat4 {
 	}}
 }
 
+// Return a Scale Matrix
+// | sx  0  0  0 |
+// |  0 sy  0  0 |
+// |  0  0 sx  0 |
+// |  0  0  0  1 |
 func Mat4MakeScale(sx, sy, sz float64) Mat4 {
 	m := Mat4Identity()
 	m.m[0][0] = sx
 	m.m[1][1] = sy
 	m.m[2][2] = sz
+	return m
+}
+
+// Return a Translation Matrix
+// | 1  0  0  tx |
+// | 0  1  0  ty |
+// | 0  0  1  tz |
+// | 0  0  0   1 |
+func Mat4MakeTrans(tx, ty, tz float64) Mat4 {
+	m := Mat4Identity()
+	m.m[0][3] = tx
+	m.m[1][3] = ty
+	m.m[2][3] = tz
 	return m
 }